chore(ui): define interfaces in place
